chore: move usePin to related domain
