+++ src/modules/blockchainApplication/manage/hooks/usePinBlockchainApplication.test.js
@@ -1,8 +1,8 @@
 import { renderHook, act } from '@testing-library/react-hooks';
 import mockBlockchainApplications from '@tests/fixtures/blockchainApplicationsExplore';
-import actionTypes from '@blockchainApplication/manage/store/actionTypes';
 import { usePinBlockchainApplication } from './usePinBlockchainApplication';
-import { toggleApplicationPin } from '../../manage/store/action';
+import actionTypes from '../store/actionTypes';
+import { toggleApplicationPin } from '../store/action';
 
 const mockDispatch = jest.fn();
 const mockState = {
